requirements.txt
python-dotenv[cli]
fabric>=2
requests
